invoice_parser-main/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "invoice-parser"
version = "0.1.0"
description = "AI-powered invoice parser with database persistence"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "uvicorn[standard]>=0.35.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.27",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "psutil>=7.0.0",
    "email-validator>=2.3.0",
    "bleach>=6.2.0",
]

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-dir]
"" = "."
